changed destinationTag parameter type

+++ contracts/assetManager/facets/RedemptionRequestsFacet.sol
@@ -93,7 +93,8 @@ contract RedemptionRequestsFacet is AssetManagerBase, ReentrancyGuard {
      *      that will be received by the redeemer - the redemption fee will be subtracted
      * @param _redeemerUnderlyingAddressString the address to which the agent must transfer underlying amount
      * @param _executor the account that is allowed to execute redemption default (besides redeemer and agent)
-     * @param _destinationTag the destination tag that is required in the redemption payment (only for XRP)
+     * @param _destinationTag the destination tag that is required in the redemption payment (only for XRP;
+     *      it must fit in 32 bits for now)
      * @return _redeemedAmountUBA the actual redeemed amount; may be less than requested if there are not enough
      *      redemption tickets available or the maximum redemption ticket limit is reached
      */
@@ -101,7 +102,7 @@ contract RedemptionRequestsFacet is AssetManagerBase, ReentrancyGuard {
         uint256 _amountUBA,
         string memory _redeemerUnderlyingAddressString,
         address payable _executor,
-        uint64 _destinationTag
+        uint256 _destinationTag
     )
         external payable
         notEmergencyPaused
@@ -112,7 +113,7 @@ contract RedemptionRequestsFacet is AssetManagerBase, ReentrancyGuard {
         uint64 amountAMG = Conversion.convertUBAToAmg(_amountUBA);
         require(amountAMG >= RedemptionRequests.minimumRedeemAmountAMG(), RedemptionTooSmall());
         require(_destinationTag < 2 ** 32, DestinationTagTooBig());
-        return _redeem(amountAMG, 1, _redeemerUnderlyingAddressString, _executor, true, _destinationTag);
+        return _redeem(amountAMG, 1, _redeemerUnderlyingAddressString, _executor, true, uint64(_destinationTag));
     }
 
     /**
